Ablitilty to simply `Sql2o.wrap` existing jdbc connection, without dealing with DataSource. `Connection` converted to interface with two implementations.
